Fix SettingWithCopyWarnings and R script call
Multiple instances of dataframe.ix references that were being processed and then
replacing the original value in the dataframe or double references that all led to
SettingWithCopyWarnings.
This reference was awesome in explaining the source of why:

https://www.dataquest.io/blog/settingwithcopywarning/

Fixed all these warnings in heatmap python script either by chaning certain instances
of .ix with .loc and other times with dataframe.copy()

Also fixed call to heatmap3_new.R which was hard coded in. Location of this R script
is now input parameter into the python script. Hacky solution that was fastest way to
this semi-responsibly.