GRmax barplots compatible with Hafner dataset
Adjusted GRmax to be more versatile to take in input from Hafner dataset with
GRserum added. For making plots of Hafner's drug response data.

- remove rows with nan GRserum values
- c is optional parameter. Input the cell line class and color. Scanr analysis
required a mapping file, cell line-subtype-color, to make the figures. Adjusted
for Hafner which has subtypes in main matrix, and defaults to a list of 9 colors.
- adjusted some other minor bugs